+++ ILPB4_OJT_FRONTEND/src/app/app.component.html
@@ -1,5 +1,3 @@
-<!-- <app-marketlist></app-marketlist> -->
-<!-- <app-sub-group></app-sub-group> -->
 <app-navbar></app-navbar>
 <router-outlet></router-outlet>
 <!-- <app-view-market-details></app-view-market-details> -->
