+++ ILPB4_OJT_FRONTEND/src/app/app.component.ts
@@ -1,9 +1,11 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { MarketlistComponent } from "./features/marketlist/marketlist.component";
-import { SubGroupComponent } from "./features/sub-group/sub-group.component";
+import { MarketlistComponent } from "./features/market-list/market-list.component";
 import { NavbarComponent } from './shared/navbar/navbar.component';
 import { ViewMarketDetailsComponent } from "./features/view-market-details/view-market-details.component";
+
+
+
 
 @Component({
   selector: 'app-root',
@@ -13,7 +15,8 @@ import { ViewMarketDetailsComponent } from "./features/view-market-details/view-
     MarketlistComponent, 
     NavbarComponent, 
     ViewMarketDetailsComponent,
-    SubGroupComponent
+   
+
   ],
   templateUrl: './app.component.html',
   styleUrl: './app.component.css'
