+++ ILPB4_OJT_FRONTEND/src/app/app.routes.ts
@@ -1,19 +1,14 @@
 import { Routes } from '@angular/router';
-import { MarketlistComponent } from './features/marketlist/marketlist.component'; // Make sure to update the path
+import { MarketlistComponent } from './features/market-list/market-list.component'; // Make sure to update the path
 import { CreateMarketComponent } from './features/create-market/create-market.component';
-import { SubGroupComponent } from './features/sub-group/sub-group.component';
 import { ViewMarketDetailsComponent } from './features/view-market-details/view-market-details.component';
 import { EditMarketComponent } from './features/edit-market/edit-market.component';
 
 export const routes: Routes = [
   {
     path: '',
-    redirectTo: '/marketlist',
+    redirectTo: 'marketlist',
     pathMatch: 'full'
-  },
-  {
-    path:'subgroup',
-    component: SubGroupComponent
   },
   {
     path: 'marketlist',
