--- ILPB4_OJT_FRONTEND/src/app/core/market-details.ts
@@ -1,11 +1,0 @@
-import { MarketSubgroup } from './market-subgroup'
-
-export interface MarketDetails {
-    marketId: number;
-    marketName: string;
-    marketCode: string;
-    longMarketCode: string;
-    region: string;
-    subRegion: string;
-    marketSubGroups: MarketSubgroup[];
-}
