--- ILPB4_OJT_FRONTEND/src/app/core/market-subgroup.ts
@@ -1,6 +1,0 @@
-export interface MarketSubgroup {
-    subGroupCode: string;
-    subGroupId: number;
-    subGroupName: string;
-
-}
