--- ILPB4_OJT_FRONTEND/src/app/core/services/market.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { MarketService } from './market.service';
-
-describe('MarketService', () => {
-  let service: MarketService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(MarketService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
