+++ ILPB4_OJT_FRONTEND/src/app/features/create-market/create-market.component.html
@@ -116,4 +116,4 @@
       </button>
     </form>
   </div>
-</div>  
+</div>
