+++ ILPB4_OJT_FRONTEND/src/app/features/edit-market/edit-market.component.ts
@@ -19,7 +19,7 @@ import { MessageService } from 'primeng/api';
 import { HeaderComponent } from '../../shared/header/header.component';
 
 import { InputMaskModule } from 'primeng/inputmask';
-import { RegionService } from '../../core/services/region.service';
+import { RegionService } from '../../services/region.service';
 
 
 /**
