--- ILPB4_OJT_FRONTEND/src/app/features/sub-group/sub-group.component.css
@@ -1,98 +1,0 @@
-
-
-  ::ng-deep .name-input {
-    width: 400px;
-    margin-right: 10px;
-  }
-
-  ::ng-deep .p-datatable-tbody > tr {
-    border-top: none;   /* Remove top border */
-    border-bottom: none; /* Remove bottom border */
-  }
-  ::ng-deep .p-datatable-tbody > tr > td {
-    border: none; /* Remove all borders for the table cells */
-  }
-
-
-  ::ng-deep .p-confirm-dialog-accept.p-button {
-    margin-left: 10px;  /* Adds space between the accept and reject buttons */
-  }
-
-
-  ::ng-deep .p-datatable-thead > tr > th,
-  ::ng-deep .p-datatable-tfoot > tr > td {
-  border: none; /* Remove borders in the header and footer */
-  }
-
-  
-  /* Adjust the delete button styling */
-  .delete-btn {
-    padding: 6px;
-  }
-  
-  /* Flexbox for inline form controls */
-  .form-inline {
-    display: flex;
-    align-items: center;
-  }
-  
-  /* General button styling for uniformity */
-  button.p-button {
-    margin: 0px 10px;
-  }
-  
-  .p-mt-3 {
-    margin-top: 20px;
-  }
-  
-
-  ::ng-deep .market-input{
-    border-right: 0;
-  }
-
-  ::ng-deep .subgroup-input{
-    border-left: 0;
-  }
-
-::ng-deep .market-input{
-    width: 50px; /* Adjust the width as per your requirement */
-}
-
-::ng-deep .subgroup-input {
-  width: 150px;
-}
-
-
-::ng-deep .p-datatable .p-datatable-tbody > tr > td {
-  padding: .5Rem;
-}
-
-::ng-deep .p-divider.p-divider-vertical {
-  margin: 0;
-  padding: 0;
-}
-
-/* Remove hover effect for PrimeNG input fields */
-::ng-deep .p-inputtext:hover, ::ng-deep .p-inputtext:focus, ::ng-deep .p-inputtext {
-  background-color: transparent !important; /* Remove background color */
-  border-color: #d1d5db !important; /* Keep border color unchanged */
-  box-shadow: none !important; /* Remove any shadow */
-  cursor: auto;
-}
-
-/* Remove focus effect for PrimeNG input fields */
-::ng-deep .p-inputtext:focus {
-  outline: none !important;
-}
-
-/* Grey background for disabled input fields */
-::ng-deep .p-inputtext:disabled,
-::ng-deep input:disabled {
-  background-color: #f0f0f0 !important; /* Set grey background */
-  color: #6c757d !important;           /* Optional: Set text color for better visibility */
-  cursor: not-allowed;                 /* Optional: Set cursor to not-allowed */
-  border-color: #d1d5db !important;    /* Keep the original border */
-  box-shadow: none !important;         /* Remove any shadow */
-}
-
-
