--- ILPB4_OJT_FRONTEND/src/app/features/sub-group/sub-group.component.html
@@ -1,73 +1,0 @@
-<p-confirmDialog></p-confirmDialog>
-
-<div class="form-container">
-  <form [formGroup]="form">
-
-
-    <p-table [value]="rows.controls" formArrayName="rows" class="p-mb-3">
-
-      <ng-template pTemplate="header">
-        <tr>
-          <th>Code</th>
-          <th>Name</th>
-        </tr>
-      </ng-template>
-
-
-      <ng-template pTemplate="body" let-row let-rowIndex="rowIndex">
-        <tr [formGroupName]="rowIndex">
-
-          <td>
-            <p-inputGroup>
-              <input type="text" placeholder="" pInputText formControlName="marketCode" maxlength="2"
-                class="market-input" readonly />
-              <p-divider layout="vertical"></p-divider>
-
-              <input type="text" placeholder="Subgroup Code" pInputText formControlName="subGroupCode" maxlength="1"
-                class="subgroup-input" (blur)="row.get('subGroupCode')?.markAsTouched()" />
-            </p-inputGroup>
-          </td>
-
-          <td>
-            <div class="form-inline">
-              <input type="text" placeholder="Subgroup Name" pInputText formControlName="subGroupName"
-                class="name-input" (blur)="row.get('subGroupName')?.markAsTouched()" />
-
-              <i style="cursor: pointer;" class="fa fa-trash-can delete-icon p-ml-2" (click)="deleteRow(rowIndex, row.get('subGroupId').value)"></i>
-            </div>
-          </td>
-        </tr>
-
-        <tr>
-          <td colspan="2">
-            <small class="p-error"
-              *ngIf="row.get('subGroupCode')?.hasError('invalidFormat') && row.get('subGroupCode')?.touched">
-              <span class="error-icon">❗</span>Invalid code format. Only a single alphanumeric character is allowed.
-            </small>
-            <small class="p-error"
-              *ngIf="row.get('subGroupCode')?.hasError('required') && row.get('subGroupCode')?.touched">
-              <span class="error-icon">❗</span>Subgroup Code is required.
-            </small>
-            <small class="p-error"
-              *ngIf="row.get('subGroupName')?.hasError('required') && row.get('subGroupName')?.touched">
-              <span class="error-icon">❗</span>Subgroup Name is required.
-            </small>
-            <small class="p-error"
-              *ngIf="row.get('subGroupCode')?.hasError('duplicateSubgroupCode') && row.get('subGroupCode')?.touched">
-              <span class="error-icon">❗</span>Subgroup Code already exists for this market.
-            </small>
-            <small class="p-error"
-              *ngIf="row.get('subGroupName')?.hasError('duplicateSubgroupName') && row.get('subGroupName')?.touched">
-              <span class="error-icon">❗</span>Subgroup Name already exists.
-            </small>
-          </td>
-        </tr>
-      </ng-template>
-    </p-table>
-
-    <button type="button" pButton icon="pi pi-plus" label="Add Subgroup" [outlined]="true" (click)="addRow()"
-      [disabled]="!canAddSubgroup()"></button>
-
-    <!-- <button type="button" pButton label="Cancel" class="p-button-danger p-mt-3" (click)="onCancel()"></button> -->
-  </form>
-</div>
