+++ ILPB4_OJT_FRONTEND/src/app/features/view-market-details/view-market-details.component.html
@@ -1,6 +1,6 @@
 <app-header [title]="marketDetails?.marketCode + ' - ' + marketDetails?.marketName">
   
-  <button class="edit-button">
+  <button class="edit-button" (click)="navigateToEdit()">
     <img src="assets/edit-icon.svg" alt="Edit Icon" width="16" height="16" />
     Edit
     <span class="notification-dot"></span>
@@ -8,7 +8,7 @@
 
 
   <button class="options-button">
-    <img src="assets/options-icon.svg" alt="Options Icon" width="16" height="16" />
+    <img src="../assets/options-icon.svg" alt="Options Icon" width="16" height="16" />
   </button>
 
 </app-header>
